feat: Supports running playbook with parameters
